requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-filebrowser==3.13.2
django-filebrowser-no-grappelli==3.8.0
django-grappelli==2.14.3
django-rest-knox==4.1.0
djangorestframework==3.12.2
environs==9.3.1
marshmallow==3.10.0
Pillow==8.2.0
pycparser==2.20
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
